Make cmake variables cached
Without caching, variables are hardcoded, so command-line customizations
are ignored, for example:

```
$ cmake -B build -DCMAKE_INSTALL_MESSAGE=ALWAYS
```

Even if `CMAKE_INSTALL_MESAGE` is set to `ALWAYS` in the command above,
it was still being set to `NEVER` since it was hardcoded in
CMakeLists-common.txt. With `CACHE` now it is set to `ALWAYS` as
expected. See:
https://cliutils.gitlab.io/modern-cmake/chapters/basics/variables.html.